// src/store/laboratory/laboratory.types.ts
export enum LABORATORY_ACTION_TYPES {
  SET_LABORATORY_START = 'laboratory/SET_LABORATORY_START',
  SET_LABORATORY_FAILED = 'laboratory/SET_LABORATORY_SUCCESS',
  SET_LABORATORY_SUCCESS = 'laboratory/SET_LABORATORY_SUCCESS',

  ADD_MEMBER_START = 'laboratory/ADD_MEMBER_START',
  ADD_MEMBER_SUCCESS = 'laboratory/ADD_MEMBER_SUCCESS',
  ADD_MEMBER_FAILED = 'laboratory/ADD_MEMBER_FAILED',

  REJECT_MEMBER_START = 'laboratory/REJECT_MEMBER_START',
  REJECT_MEMBER_SUCCESS = 'laboratory/REJECT_MEMBER_SUCCESS',
  REJECT_MEMBER_FAILED = 'laboratory/REJECT_MEMBER_FAILED',

  SET_ROLE_MEMBER_START = 'laboratory/SET_ROLE_MEMBER_START',
  SET_ROLE_MEMBER_SUCCESS = 'laboratory/SET_ROLE_MEMBER_SUCCESS',
  SET_ROLE_MEMBER_FAILED = 'laboratory/SET_ROLE_MEMBER_FAILED',

  REMOVE_MEMBER_START = 'laboratory/REMOVE_MEMBER_START',
  REMOVE_MEMBER_SUCCESS = 'laboratory/REMOVE_MEMBER_SUCCESS',
  REMOVE_MEMBER_FAILED = 'laboratory/REMOVE_MEMBER_FAILED',

  ACCEPT_RESEARCH_START = 'laboratory/ACCEPT_RESEARCH_START',
  ACCEPT_RESEARCH_SUCCESS = 'laboratory/ACCEPT_RESEARCH_SUCCESS',
  ACCEPT_RESEARCH_FAILED = 'laboratory/ACCEPT_RESEARCH_FAILED',

  REJECT_RESEARCH_START = 'laboratory/REJECT_RESEARCH_START',
  REJECT_RESEARCH_SUCCESS = 'laboratory/REJECT_RESEARCH_SUCCESS',
  REJECT_RESEARCH_FAILED = 'laboratory/REJECT_RESEARCH_FAILED',

  GET_MEMBER_WAIT_START = 'laboratory/GET_MEMBER_WAIT_START',
  GET_MEMBER_WAIT_SUCCESS = 'laboratory/GET_MEMBER_WAIT_SUCCESS',
  GET_MEMBER_WAIT_FAILED = 'laboratory/GET_MEMBER_WAIT_FAILED',

  GET_RESEARCH_WAIT_START = 'laboratory/GET_RESEARCH_WAIT_START',
  GET_RESEARCH_WAIT_SUCCESS = 'laboratory/GET_RESEARCH_WAIT_SUCCESS',
  GET_RESEARCH_WAIT_FAILED = 'laboratory/GET_RESEARCH_WAIT_FAILED'
}